requirements.txt
Flask==2.3.2
instagrapi==1.16.20
python-dotenv==0.19.0
requests==2.31.0